requirements.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord==1.0.1
discord.py==1.3.4
Django==3.0.8
django-rest-framework==0.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
idna==2.10
multidict==4.7.6
NBT==1.5.0
numpy==1.19.0
pandas==1.0.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
